Split spiffe verifier to its own package
